requirements.txt
# Core Dependencies
numpy>=1.23.0
opencv-python>=4.8.0
Pillow>=9.5.0
ultralytics>=8.0.0
supervision>=0.3.0
streamlit>=1.27.0

# Optional utilities
argparse>=1.4.0
xmltodict>=0.13.0
